separate ios/android logic
