+++ src/index.d.ts
@@ -1,3 +1,5 @@
 export { ModalStack } from "./modal-stack";
 export { default as VueWindowedModal } from "./vue-windowed-modal";
-export { overrideModalViewMethod } from "./windowed-modal.common";
+export { overrideModalViewMethod } from "./windowed-modal";
+export { ExtendedShowModalOptions } from "./windowed-modal.common";
+
