+++ src/windowed-modal.d.ts
@@ -1,0 +1,2 @@
+
+export function overrideModalViewMethod(): void;
